add symbolic base class
